requirements.txt
ascii-animator==0.1.8
ascii-magic==1.6
asciimatics==1.15.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
cursor==1.3.5
idna==3.7
list2term==0.1.6
pillow==10.3.0
prompt_toolkit==3.0.47
pyfiglet==1.0.2
requests==2.32.3
urllib3==2.2.2
wcwidth==0.2.13
